carbon/src/ImageTools/ImageTools_v141.vcxproj: Undoing "TEMP" hack (from #3) to restore linkage with "ImageIO.lib" (instead of "ImageIO_dev.lib") in the "Release|x64" project config.
This should fix the "Deploy_D3D11" solution config, at the expense of breaking "Dev_D3D11" again (but the former is more important).

Ultimately, the Dev solution still needs to be fixed "properly" by creating a project-level "Dev" config for ImageTools and setting the Dev sln config to use that one.. assuming anyone actually cares about this crap (which they possibly don't).

BTW: A "Dev" config is supposedly just an "Unoptimzed Release" config, and these are used in a few places as a more practical substitute for the essentially useless "Debug" configs (which are always too slow, and broken, to be useful).

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/ImageTools/": change = 1545490]