Image Server Icon Validation --  https://jira.ccpgames.com/browse/TC-247
Adding the validation script for the image server icon generation (still WIP).
Moved the confluence.py utility file to a more generic location so it can be shared by both the Graphics Report and the Icon Validation.
Changed the way RMS tolerance works in the TextureDiff package to make it dependent on the size of the images being compared. Also fixed the texel tolerance, which didn't really work as intended.

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/packages/imageutils/": change = 1632265]