Add find package for NanoVDB and fix missing TrinityAL includes (#1)
ImageIO now uses NanoVDB so this project required a find_package to
locate it.

Project uses some headers from TrinityAL, as it was not linked the headers
were not automatically included in the project so added a line to include
the headers from TrinityAL_stub. This is OK as this project in Windows only.