Adding missing files to image tools
Apparently we were missing a few files for carbon imagetools.
Adding them and massaging them to fit with the githubness of it all.
For example the NanoVDBToolSupport files were supposed to be called NanoVDBSupport but the name clashed with files named NanoVDBSupport from ImageIO which this project is including.
So there is maybe a bigger question on how these things are supposed to work, in a post github world...